backend/requirements.txt
# File: backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Pillow
requests
beautifulsoup4
lxml
python-dateutil
python-dotenv
gunicorn
dj-database-url
psycopg2-binary
whitenoise
